debug post send and object_manager object in object_manager_rdma

+++ src/ray/object_manager/object_manager_rdma.cc
@@ -448,7 +448,7 @@ void ObjectManagerRdma::FetchObjectFromRemotePlasma(const ray::WorkerID &worker_
       std::ofstream outfile;
       std::string filename = "buffer.txt";
       void *buffer = (void *) local_address;
-      uint8_t *buf = (uint8_t*) buffer;
+      uint64_t *buf = (uint64_t*) buffer;
       outfile.open(filename);
       for(int i=0; i<object_sizes[i]; ++i){
         outfile<<buf[i];
